+++ examples/advanced/src/gen/clients/axios/index.ts
@@ -1,0 +1,4 @@
+export * as petService from './petService'
+export * as storeService from './storeService'
+export * as userService from './userService'
+export * from './operations'