+++ examples/advanced/src/gen/clients/hooks/index.ts
@@ -1,0 +1,3 @@
+export * as petHooks from './petController'
+export * as storeHooks from './storeController'
+export * as userHooks from './userController'