+++ examples/advanced/src/gen/clients/hooks/petController/useAddPet.ts
@@ -1,0 +1,28 @@
+import { useMutation } from '@tanstack/react-query'
+
+import client from '../../../../client'
+
+import type { UseMutationOptions } from '@tanstack/react-query'
+import type { AddPetRequest, AddPetResponse, AddPet405 } from '../../../models/ts/petController/AddPet'
+
+/**
+ * @description Add a new pet to the store
+ * @summary Add a new pet to the store
+ * @link /pet
+ */
+export function useAddPet<TData = AddPetResponse, TError = AddPet405, TVariables = AddPetRequest>(options?: {
+  mutation?: UseMutationOptions<TData, TError, TVariables>
+}) {
+  const { mutation: mutationOptions } = options ?? {}
+
+  return useMutation<TData, TError, TVariables>({
+    mutationFn: (data) => {
+      return client<TData, TVariables>({
+        method: 'post',
+        url: `/pet`,
+        data,
+      })
+    },
+    ...mutationOptions,
+  })
+}