+++ examples/advanced/src/gen/clients/hooks/userController/useCreateUser.ts
@@ -1,0 +1,28 @@
+import { useMutation } from '@tanstack/react-query'
+
+import client from '../../../../client'
+
+import type { UseMutationOptions } from '@tanstack/react-query'
+import type { CreateUserRequest, CreateUserResponse } from '../../../models/ts/userController/CreateUser'
+
+/**
+ * @description This can only be done by the logged in user.
+ * @summary Create user
+ * @link /user
+ */
+export function useCreateUser<TData = CreateUserResponse, TError = unknown, TVariables = CreateUserRequest>(options?: {
+  mutation?: UseMutationOptions<TData, TError, TVariables>
+}) {
+  const { mutation: mutationOptions } = options ?? {}
+
+  return useMutation<TData, TError, TVariables>({
+    mutationFn: (data) => {
+      return client<TData, TVariables>({
+        method: 'post',
+        url: `/user`,
+        data,
+      })
+    },
+    ...mutationOptions,
+  })
+}