+++ examples/advanced/src/gen/clients/hooks/userController/useLogoutUser.ts
@@ -1,0 +1,42 @@
+import { useQuery } from '@tanstack/react-query'
+
+import client from '../../../../client'
+
+import type { QueryKey, UseQueryResult, UseQueryOptions, QueryOptions } from '@tanstack/react-query'
+import type { LogoutUserResponse } from '../../../models/ts/userController/LogoutUser'
+
+export const logoutUserQueryKey = () => [`/user/logout`] as const
+
+export function logoutUserQueryOptions<TData = LogoutUserResponse>(): QueryOptions<TData> {
+  const queryKey = logoutUserQueryKey()
+
+  return {
+    queryKey,
+    queryFn: () => {
+      return client<TData>({
+        method: 'get',
+        url: `/user/logout`,
+      })
+    },
+  }
+}
+
+/**
+ * @summary Logs out current logged in user session
+ * @link /user/logout
+ */
+export function useLogoutUser<TData = LogoutUserResponse, TError = unknown>(options?: {
+  query?: UseQueryOptions<TData>
+}): UseQueryResult<TData, unknown> & { queryKey: QueryKey } {
+  const { query: queryOptions } = options ?? {}
+  const queryKey = queryOptions?.queryKey ?? logoutUserQueryKey()
+
+  const query = useQuery<TData, TError>({
+    ...logoutUserQueryOptions<TData>(),
+    ...queryOptions,
+  }) as UseQueryResult<TData, TError> & { queryKey: QueryKey }
+
+  query.queryKey = queryKey as QueryKey
+
+  return query
+}