+++ examples/advanced/src/gen/reactQuery/index.ts
@@ -1,3 +1,0 @@
-export * from './petController'
-export * from './storeController'
-export * from './userController'